+++ featurebyte/routes/event_table/controller.py
@@ -13,6 +13,10 @@ from featurebyte.routes.common.base_table import BaseTableDocumentController
 from featurebyte.schema.event_table import EventTableList, EventTableServiceUpdate
 from featurebyte.schema.info import EventTableInfo
 from featurebyte.service.event_table import EventTableService
+from featurebyte.service.semantic import SemanticService
+from featurebyte.service.table_columns_info import TableColumnsInfoService, TableDocumentService
+from featurebyte.service.table_info import TableInfoService
+from featurebyte.service.table_status import TableStatusService
 
 
 class EventTableController(
@@ -24,6 +28,19 @@ class EventTableController(
 
     paginated_document_class = EventTableList
     document_update_schema_class = EventTableServiceUpdate
+
+    def __init__(
+        self,
+        service: TableDocumentService,
+        table_columns_info_service: TableColumnsInfoService,
+        table_status_service: TableStatusService,
+        semantic_service: SemanticService,
+        table_info_service: TableInfoService,
+    ):
+        super().__init__(
+            service, table_columns_info_service, table_status_service, semantic_service
+        )
+        self.table_info_service = table_info_service
 
     async def _get_column_semantic_map(self, document: EventTableModel) -> dict[str, Any]:
         event_timestamp = await self.semantic_service.get_or_create_document(
@@ -51,7 +68,13 @@ class EventTableController(
         -------
         EventTableInfo
         """
-        info_document = await self.service.get_event_table_info(
-            document_id=document_id, verbose=verbose
+        event_table = await self.service.get_document(document_id=document_id)
+        table_dict = await self.table_info_service.get_table_info(
+            data_document=event_table, verbose=verbose
         )
-        return info_document
+        return EventTableInfo(
+            **table_dict,
+            event_id_column=event_table.event_id_column,
+            event_timestamp_column=event_table.event_timestamp_column,
+            default_feature_job_setting=event_table.default_feature_job_setting,
+        )
