+++ featurebyte/routes/scd_table/controller.py
@@ -13,6 +13,10 @@ from featurebyte.routes.common.base_table import BaseTableDocumentController
 from featurebyte.schema.info import SCDTableInfo
 from featurebyte.schema.scd_table import SCDTableList, SCDTableServiceUpdate
 from featurebyte.service.scd_table import SCDTableService
+from featurebyte.service.semantic import SemanticService
+from featurebyte.service.table_columns_info import TableColumnsInfoService, TableDocumentService
+from featurebyte.service.table_info import TableInfoService
+from featurebyte.service.table_status import TableStatusService
 
 
 class SCDTableController(BaseTableDocumentController[SCDTableModel, SCDTableService, SCDTableList]):
@@ -22,6 +26,19 @@ class SCDTableController(BaseTableDocumentController[SCDTableModel, SCDTableServ
 
     paginated_document_class = SCDTableList
     document_update_schema_class = SCDTableServiceUpdate
+
+    def __init__(
+        self,
+        service: TableDocumentService,
+        table_columns_info_service: TableColumnsInfoService,
+        table_status_service: TableStatusService,
+        semantic_service: SemanticService,
+        table_info_service: TableInfoService,
+    ):
+        super().__init__(
+            service, table_columns_info_service, table_status_service, semantic_service
+        )
+        self.table_info_service = table_info_service
 
     async def _get_column_semantic_map(self, document: SCDTableModel) -> dict[str, Any]:
         scd_natural_key_id = await self.semantic_service.get_or_create_document(
@@ -56,7 +73,15 @@ class SCDTableController(BaseTableDocumentController[SCDTableModel, SCDTableServ
         -------
         SCDTableInfo
         """
-        info_document = await self.service.get_scd_table_info(
-            document_id=document_id, verbose=verbose
+        scd_table = await self.service.get_document(document_id=document_id)
+        table_dict = await self.table_info_service.get_table_info(
+            data_document=scd_table, verbose=verbose
         )
-        return info_document
+        return SCDTableInfo(
+            **table_dict,
+            natural_key_column=scd_table.natural_key_column,
+            effective_timestamp_column=scd_table.effective_timestamp_column,
+            surrogate_key_column=scd_table.surrogate_key_column,
+            end_timestamp_column=scd_table.end_timestamp_column,
+            current_flag_column=scd_table.current_flag_column,
+        )
