+++ featurebyte/routes/static_source_table/controller.py
@@ -11,6 +11,7 @@ from featurebyte.routes.task.controller import TaskController
 from featurebyte.schema.info import StaticSourceTableInfo
 from featurebyte.schema.static_source_table import StaticSourceTableCreate, StaticSourceTableList
 from featurebyte.schema.task import Task
+from featurebyte.service.feature_store import FeatureStoreService
 from featurebyte.service.preview import PreviewService
 from featurebyte.service.static_source_table import StaticSourceTableService
 from featurebyte.service.table import TableService
@@ -34,10 +35,12 @@ class StaticSourceTableController(
         preview_service: PreviewService,
         table_service: TableService,
         task_controller: TaskController,
+        feature_store_service: FeatureStoreService,
     ):
         super().__init__(service=service, preview_service=preview_service)
         self.table_service = table_service
         self.task_controller = task_controller
+        self.feature_store_service = feature_store_service
 
     async def create_static_source_table(
         self,
@@ -81,7 +84,16 @@ class StaticSourceTableController(
         -------
         StaticSourceTableInfo
         """
-        info_document = await self.service.get_static_source_table_info(
-            document_id=document_id, verbose=verbose
+        _ = verbose
+        static_source_table = await self.service.get_document(document_id=document_id)
+        feature_store = await self.feature_store_service.get_document(
+            document_id=static_source_table.location.feature_store_id
         )
-        return info_document
+        return StaticSourceTableInfo(
+            name=static_source_table.name,
+            type=static_source_table.request_input.type,
+            feature_store_name=feature_store.name,
+            table_details=static_source_table.location.table_details,
+            created_at=static_source_table.created_at,
+            updated_at=static_source_table.updated_at,
+        )
