+++ featurebyte/service/historical_feature_table.py
@@ -3,7 +3,7 @@ HistoricalFeatureTableService class
 """
 from __future__ import annotations
 
-from typing import Any, Optional
+from typing import Optional
 
 from pathlib import Path
 
@@ -13,16 +13,11 @@ from bson import ObjectId
 from featurebyte.enum import MaterializedTableNamePrefix
 from featurebyte.models.base import FeatureByteBaseDocumentModel
 from featurebyte.models.historical_feature_table import HistoricalFeatureTableModel
-from featurebyte.persistent import Persistent
 from featurebyte.schema.historical_feature_table import HistoricalFeatureTableCreate
-from featurebyte.schema.info import HistoricalFeatureTableInfo
 from featurebyte.schema.worker.task.historical_feature_table import (
     HistoricalFeatureTableTaskPayload,
 )
-from featurebyte.service.feature_list import FeatureListService
-from featurebyte.service.feature_store import FeatureStoreService
 from featurebyte.service.materialized_table import BaseMaterializedTableService
-from featurebyte.service.observation_table import ObservationTableService
 from featurebyte.storage import Storage
 
 
@@ -35,19 +30,6 @@ class HistoricalFeatureTableService(
 
     document_class = HistoricalFeatureTableModel
     materialized_table_name_prefix = MaterializedTableNamePrefix.HISTORICAL_FEATURE_TABLE
-
-    def __init__(
-        self,
-        user: Any,
-        persistent: Persistent,
-        catalog_id: ObjectId,
-        feature_store_service: FeatureStoreService,
-        observation_table_service: ObservationTableService,
-        feature_list_service: FeatureListService,
-    ):
-        super().__init__(user, persistent, catalog_id, feature_store_service)
-        self.observation_table_service = observation_table_service
-        self.feature_list_service = feature_list_service
 
     @property
     def class_name(self) -> str:
@@ -100,42 +82,4 @@ class HistoricalFeatureTableService(
             catalog_id=self.catalog_id,
             output_document_id=output_document_id,
             observation_set_storage_path=observation_set_storage_path,
-        )
-
-    async def get_historical_feature_table_info(
-        self, document_id: ObjectId, verbose: bool
-    ) -> HistoricalFeatureTableInfo:
-        """
-        Get historical feature table info
-
-        Parameters
-        ----------
-        document_id: ObjectId
-            Document ID
-        verbose: bool
-            Verbose or not
-
-        Returns
-        -------
-        HistoricalFeatureTableInfo
-        """
-        _ = verbose
-        historical_feature_table = await self.get_document(document_id=document_id)
-        if historical_feature_table.observation_table_id is not None:
-            observation_table = await self.observation_table_service.get_document(
-                document_id=historical_feature_table.observation_table_id
-            )
-        else:
-            observation_table = None
-        feature_list = await self.feature_list_service.get_document(
-            document_id=historical_feature_table.feature_list_id
-        )
-        return HistoricalFeatureTableInfo(
-            name=historical_feature_table.name,
-            feature_list_name=feature_list.name,
-            feature_list_version=feature_list.version.to_str(),
-            observation_table_name=observation_table.name if observation_table else None,
-            table_details=historical_feature_table.location.table_details,
-            created_at=historical_feature_table.created_at,
-            updated_at=historical_feature_table.updated_at,
         )
