+++ featurebyte/service/static_source_table.py
@@ -9,12 +9,9 @@ from bson import ObjectId
 
 from featurebyte.models.base import FeatureByteBaseDocumentModel
 from featurebyte.models.static_source_table import StaticSourceTableModel
-from featurebyte.persistent import Persistent
 from featurebyte.query_graph.node.schema import TableDetails
-from featurebyte.schema.info import StaticSourceTableInfo
 from featurebyte.schema.static_source_table import StaticSourceTableCreate
 from featurebyte.schema.worker.task.static_source_table import StaticSourceTableTaskPayload
-from featurebyte.service.feature_store import FeatureStoreService
 from featurebyte.service.materialized_table import BaseMaterializedTableService
 from featurebyte.session.base import BaseSession
 
@@ -28,15 +25,6 @@ class StaticSourceTableService(
 
     document_class = StaticSourceTableModel
     materialized_table_name_prefix = "STATIC_SOURCE_TABLE"
-
-    def __init__(
-        self,
-        user: Any,
-        persistent: Persistent,
-        catalog_id: ObjectId,
-        feature_store_service: FeatureStoreService,
-    ):
-        super().__init__(user, persistent, catalog_id, feature_store_service)
 
     @property
     def class_name(self) -> str:
@@ -101,34 +89,3 @@ class StaticSourceTableService(
             "columns_info": columns_info,
             "num_rows": num_rows,
         }
-
-    async def get_static_source_table_info(
-        self, document_id: ObjectId, verbose: bool
-    ) -> StaticSourceTableInfo:
-        """
-        Get static source table info
-
-        Parameters
-        ----------
-        document_id: ObjectId
-            Document ID
-        verbose: bool
-            Verbose or not
-
-        Returns
-        -------
-        StaticSourceTableInfo
-        """
-        _ = verbose
-        static_source_table = await self.get_document(document_id=document_id)
-        feature_store = await self.feature_store_service.get_document(
-            document_id=static_source_table.location.feature_store_id
-        )
-        return StaticSourceTableInfo(
-            name=static_source_table.name,
-            type=static_source_table.request_input.type,
-            feature_store_name=feature_store.name,
-            table_details=static_source_table.location.table_details,
-            created_at=static_source_table.created_at,
-            updated_at=static_source_table.updated_at,
-        )
